create rao-commons module and move rao inputs methods to this module
