+++ data/crac/crac-api/src/main/java/com/farao_community/farao/data/crac_api/Crac.java
@@ -267,7 +267,4 @@ public interface Crac extends Identifiable<Crac>, Synchronizable, NetworkElement
      * @return null if the NetworkAction does not exist in the Crac, the NetworkAction otherwise
      */
     NetworkAction getNetworkAction(String id);
-
-    // General methods
-    void generateValidityReport(Network network);
 }
