+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/SimpleCrac.java
@@ -460,58 +460,4 @@ public class SimpleCrac extends AbstractIdentifiable<Crac> implements Crac {
     public boolean isSynchronized() {
         return isSynchronized;
     }
-
-    @Override
-    public void generateValidityReport(Network network) {
-        ArrayList<Cnec> absentFromNetworkCnecs = new ArrayList<>();
-        getCnecs().forEach(cnec -> {
-            if (network.getBranch(cnec.getNetworkElement().getId()) == null) {
-                absentFromNetworkCnecs.add(cnec);
-                LOGGER.warn(String.format("Cnec %s with network element [%s] is not present in the network. It is removed from the Crac", cnec.getId(), cnec.getNetworkElement().getId()));
-            }
-        });
-        absentFromNetworkCnecs.forEach(cnec -> cnecs.remove(cnec));
-        ArrayList<RangeAction> absentFromNetworkRangeActions = new ArrayList<>();
-        for (RangeAction rangeAction: getRangeActions()) {
-            rangeAction.getNetworkElements().forEach(networkElement -> {
-                if (network.getIdentifiable(networkElement.getId()) == null) {
-                    absentFromNetworkRangeActions.add(rangeAction);
-                    LOGGER.warn(String.format("Remedial Action %s with network element [%s] is not present in the network. It is removed from the Crac", rangeAction.getId(), networkElement.getId()));
-                    return;
-                }
-            });
-        }
-        absentFromNetworkRangeActions.forEach(rangeAction -> rangeActions.remove(rangeAction));
-
-        ArrayList<NetworkAction> absentFromNetworkNetworkActions = new ArrayList<>();
-        for (NetworkAction networkAction: getNetworkActions()) {
-            networkAction.getNetworkElements().forEach(networkElement -> {
-                if (network.getIdentifiable(networkElement.getId()) == null) {
-                    absentFromNetworkNetworkActions.add(networkAction);
-                    LOGGER.warn(String.format("Remedial Action %s with network element [%s] is not present in the network. It is removed from the Crac", networkAction.getId(), networkElement.getId()));
-                    return;
-                }
-            });
-        }
-        absentFromNetworkNetworkActions.forEach(networkAction -> networkActions.remove(networkAction));
-
-        ArrayList<Contingency> absentFromNetworkContingencies = new ArrayList<>();
-        for (Contingency contingency : getContingencies()) {
-            contingency.getNetworkElements().forEach(networkElement -> {
-                if (network.getIdentifiable(networkElement.getId()) == null) {
-                    absentFromNetworkContingencies.add(contingency);
-                    LOGGER.warn(String.format("Contingency %s with network element [%s] is not present in the network. It is removed from the Crac", contingency.getId(), networkElement.getId()));
-                    return;
-                }
-            });
-        }
-
-        absentFromNetworkContingencies.forEach(contingency ->  {
-            getStatesFromContingency(contingency.getId()).forEach(state -> {
-                getCnecs(state).forEach(cnec -> this.cnecs.remove(cnec));
-                this.states.remove(state);
-            });
-            this.contingencies.remove(contingency);
-        });
-    }
 }
