+++ ra-optimisation/linear-rao/src/main/java/com/farao_community/farao/linear_rao/LinearRao.java
@@ -16,6 +16,7 @@ import com.farao_community.farao.linear_rao.config.LinearRaoParameters;
 import com.farao_community.farao.linear_rao.optimisation.LinearOptimisationException;
 import com.farao_community.farao.rao_api.RaoParameters;
 import com.farao_community.farao.rao_api.RaoProvider;
+import com.farao_community.farao.rao_commons.RaoInput;
 import com.farao_community.farao.util.NativeLibraryLoader;
 import com.farao_community.farao.rao_api.RaoResult;
 import com.farao_community.farao.util.SensitivityComputationException;
@@ -57,6 +58,8 @@ public class LinearRao implements RaoProvider {
                                              ComputationManager computationManager,
                                              RaoParameters raoParameters) {
         network.getVariantManager().setWorkingVariant(variantId);
+        RaoInput.synchronize(crac, network);
+        RaoInput.cleanCrac(crac, network);
         LinearRaoData linearRaoData = new LinearRaoData(network, crac);
         try {
             // check config
