+++ ra-optimisation/pom.xml
@@ -21,6 +21,7 @@
         <module>ra-optimisation-api</module>
         <module>rao-api</module>
         <module>search-tree-rao</module>
+        <module>rao-commons</module>
     </modules>
 
 </project>
