+++ ra-optimisation/rao-api/src/main/java/com/farao_community/farao/rao_api/RaoTool.java
@@ -110,8 +110,6 @@ public class RaoTool implements Tool {
         //Crac
         context.getOutputStream().println("Importing crac '" + cracFile + "'");
         Crac crac = CracImporters.importCrac(cracFile);
-        crac.generateValidityReport(network);
-        crac.synchronize(network);
 
         //Rao Parameter
         context.getOutputStream().println("Loading RAO parameters");
