+++ ra-optimisation/search-tree-rao/pom.xml
@@ -41,6 +41,11 @@
             <artifactId>farao-rao-api</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>farao-rao-commons</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!-- test -->
         <dependency>
@@ -115,5 +120,4 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
-
 </project>
